requirements.txt
aiogram>=3.0.0
python-dotenv>=1.0.0
pytesseract>=0.3.10
Pillow>=10.0.0
pymongo>=4.5.0
google-cloud-vision>=3.4.0
openai>=1.3.0
google-generativeai>=0.3.0
pdfplumber>=0.10.0
reportlab>=4.0.0
requests>=2.31.0
python-dateutil>=2.8.2
motor>=3.3.0  # Async MongoDB driver
aiofiles>=23.2.0  # Async file operations
pytz>=2023.3
